release: Lilo 0.1.1 cross-platform MVP
